Migrated the code to use Redis as the backing store
SQLite is great, but not a good match for async javascript
code causing concurrent writes.  Also moving to redis as it's
in-memory so reduces writes to the RPi SD Card.  If the list
of torrents is ephemeral in Redis (eg, a reboot and the data
can be lost) then the same is true for Transmission is it (rather
oddly) regenerates all the torrent id's when it restarts.

Performance has also had about a 50% boost.