Attempt fix for disconnect between localstorage & redis
When torrents are marked as copying, there can be a disconnect
between what the browser localstorage thinks the world is like, and
how redis sees it.  So if you refresh the page quickly after starting
a torrent copying it can look like it's in the wrong state.

Have put in a dirty/hacky work-around for now which does a full refresh
of the torrent list from redis and pushes it to localstorage after kicking
off the copy ajax request.

Might help...